Modernize for new libs/compiler, and remove some obsolete hacks

+++ graphic.h
@@ -83,7 +83,7 @@ class Graphic {
   Graphic OffHashed(color); 
   Graphic DoubleX();
   Graphic DoubleY();
-  long xsize, ysize, depth;
+  int xsize, ysize, depth;
   int xcenter, ycenter;
   mfmt * image;
   unsigned long tcolor;
